feat: Add scanning functionality
