+++ ui/global.css
@@ -3,6 +3,7 @@ body {
     position: relative;
     width: 100%;
     height: 100%;
+    background: #e6eaec;
 }
 
 body {
